tests/wpt/tests/encoding-detection/ru-IBM866.tentative.html

<!doctype html>
<title>ru IBM866</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>Это тест кодировки символов.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "IBM866", 'Expected IBM866');
    }, "Check detection result");
      done();
};
</script>
